look_mode and cursor control
by pressing 'k' or 'l', the cursor is shown and follows a 'cursor' tile which will eventually be assosicated with a short text description of the contents of that tile for user clarity. When the cursor is visible control is taken from player movement into a 'looking mode' where the cursor is being controlled with the arrow keys.